refactor: make seed be a uint64_t (#460)

+++ src/types/tournament_options.hpp
@@ -97,7 +97,7 @@ struct Tournament {
     OutputType output = OutputType::FASTCHESS;
 #endif
 
-    uint32_t seed = 951356066;
+    uint64_t seed = 951356066;
 
     int ratinginterval   = 10;
     int scoreinterval    = 1;
